refactor: extract parseCsv into its own module; drop fixture comment
